Export an operation that looks update up on the context
Rather than having the spinner and loader look up the update
function and then invoke it, this exports an `update()` operation that
does the lookup for you, and then invokes it. That way, you just say:

```js
yield* update(state);
```
from anywhere in one step. If the context is missing, it will
complain about it. This is the more general pattern of contextual
effects.

At the same time, this "lifts" the raw `setState()` function into an operation
which is safer. In general, any function that could potentially cause your
operation to exit should be lifted into an operation itself. That way, let's
say you call `setState()` and it causes your component to be torn
down, which then halts your effect, then you don't want anything to
run after that in your operation.